refactor(model): rename and restructure model modules
